@@ -2,9 +2,10 @@
   "name": "holochain-docs-pages",
   "version": "1.0.0",
   "description": "11ty project for generating the Holochain documentation",
-  "main": "index.js",
+  "main": "index.cjs",
+  "type": "module",
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   },
   "directories": {
     "doc": "docs"
@@ -19,9 +20,9 @@
     "watch:sass": "npx sass --watch src/scss:styles-compiled",
     "post:css": "npx postcss styles-compiled/site.css > styles-compiled/site.min.css",
     "min:js": "npx uglify-js _site/scripts/main.bundled.js --output _site/scripts/main.min.js",
-    "bundle:js": "rollup -c rollup.config.mjs",
-    "watch:js": "rollup -c rollup.config.mjs -w",
-    "dev": "npm run build:sass && npm-run-all -p serve:web watch:sass watch:js",
+    "bundle:js": "rollup -c rollup.config.js",
+    "watch:js": "rollup -c rollup.config.js -w",
+    "dev": "npm run build && npm-run-all -p serve:web watch:sass watch:js",
     "dev-no-js": "npm run build:sass && npm-run-all -p serve:web watch:sass",
     "build": "npm-run-all -s clean:site build:sass post:css clean:js bundle:js build:web build:search-index",
     "serve-as-prod": "npm-run-all -s clean:site build:sass post:css clean:js bundle:js serve:web-prod",
@@ -42,8 +43,8 @@
   },
   "homepage": "https://github.com/holochain/docs-pages#readme",
   "devDependencies": {
-    "@11ty/eleventy": "^2.0.0",
-    "@aloskutov/eleventy-plugin-external-links": "^2.0.1",
+    "@11ty/eleventy": "^3.0.0-alpha.1",
+    "@aloskutov/eleventy-plugin-external-links": "^2.1.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^11.1.5",
     "@sindresorhus/slugify": "1.1.2",
